separate out workflow information
