+++ client/src/components/Workflow/Published/WorkflowInformation.vue
@@ -1,0 +1,113 @@
+<script setup lang="ts">
+import { library } from "@fortawesome/fontawesome-svg-core";
+import { faBuilding, faUser } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
+import { computed } from "vue";
+
+import Heading from "@/components/Common/Heading.vue";
+import License from "@/components/License/License.vue";
+import StatelessTags from "@/components/TagsMultiselect/StatelessTags.vue";
+import UtcDate from "@/components/UtcDate.vue";
+
+library.add(faBuilding, faUser);
+
+interface WorkflowInformation {
+    name: string;
+    [key: string]: unknown;
+    license?: string;
+    tags?: string[];
+    update_time: string;
+    creator?: Array<{
+        [key: string]: unknown;
+    }>;
+}
+
+const props = defineProps<{
+    workflowInfo: WorkflowInformation;
+}>();
+
+const gravatarSource = computed(
+    () => `https://secure.gravatar.com/avatar/${props.workflowInfo?.email_hash}?d=identicon`
+);
+
+const publishedByUser = computed(() => `/workflows/list_published?f-username=${props.workflowInfo?.owner}`);
+</script>
+
+<template>
+    <aside class="workflow-information">
+        <hgroup>
+            <Heading h2 size="xl" class="mb-0">About This Workflow</Heading>
+            <span class="ml-2">{{ workflowInfo.name }} - Version {{ workflowInfo.version }}</span>
+        </hgroup>
+
+        <div class="workflow-info-box">
+            <hgroup class="mb-2">
+                <Heading h3 size="md" class="mb-0">Author</Heading>
+                <span class="ml-2">{{ workflowInfo.owner }}</span>
+            </hgroup>
+
+            <img alt="User Avatar" :src="gravatarSource" class="mb-2" />
+
+            <router-link :to="publishedByUser"> All published Workflows by {{ workflowInfo.owner }} </router-link>
+        </div>
+
+        <div v-if="workflowInfo?.creator" class="workflow-info-box">
+            <Heading h3 size="md" class="mb-0">Creators</Heading>
+
+            <ul class="list-unstyled mb-0">
+                <li v-for="(creator, index) in workflowInfo.creator" :key="index">
+                    <FontAwesomeIcon v-if="creator.class === 'Person'" icon="fa-user" />
+                    <FontAwesomeIcon v-if="creator.class === 'Organization'" icon="fa-building" />
+                    {{ creator.name }}
+                </li>
+            </ul>
+        </div>
+
+        <div class="workflow-info-box">
+            <Heading h3 size="md" class="mb-0">Description</Heading>
+
+            <p v-if="workflowInfo.annotation" class="mb-0">
+                {{ workflowInfo.annotation }}
+            </p>
+            <p v-else class="mb-0">This Workflow has no description.</p>
+        </div>
+
+        <div v-if="workflowInfo?.tags" class="workflow-info-box">
+            <Heading h3 size="md" class="mb-0">Tags</Heading>
+
+            <StatelessTags class="tags mt-2" :value="workflowInfo.tags" disabled />
+        </div>
+
+        <div class="workflow-info-box">
+            <Heading h3 size="md" class="mb-0">License</Heading>
+
+            <License v-if="workflowInfo.license" :license-id="workflowInfo.license" />
+            <span v-else>No License specified</span>
+        </div>
+
+        <div class="workflow-info-box">
+            <Heading h3 size="md" class="mb-0">Last Updated</Heading>
+
+            <UtcDate :date="workflowInfo.update_time" mode="pretty" />
+        </div>
+    </aside>
+</template>
+
+<style scoped lang="scss">
+.workflow-information {
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    align-items: start;
+    justify-content: start;
+    align-self: flex-start;
+    overflow-y: scroll;
+
+    .workflow-info-box {
+        display: flex;
+        flex-direction: column;
+        align-items: start;
+    }
+}
+</style>
